actualizarUsuario/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
PyJWT
python-dotenv
psycopg2-binary
flask-jwt-extended